+++ module/sheets/torgeternityItemSheet.js
@@ -101,6 +101,8 @@ export default class TorgeternityItemSheet extends ItemSheet {
 
     data.description = await TextEditor.enrichHTML(this.object.system.description, { async: true });
 
+    data.ammunition = this.document.actor?.itemTypes?.ammunition ?? [];
+
     return data;
   }
 
@@ -147,5 +149,100 @@ export default class TorgeternityItemSheet extends ItemSheet {
       const newShown = 0 < currentShown ? currentShown - 1 : currentShown;
       this.item.update({ 'system.pulpPowers.limitationNumber': newShown });
     });
+
+    html.find('.reload-weapon').click((ev) => {
+      const button = ev.currentTarget.closest('[data-item-id]');
+      const usedAmmo = this?.actor.items.get(button.dataset.itemId);
+
+      reloadAmmo(this.actor, this.object, usedAmmo);
+    });
   }
 }
+
+/**
+ * Reload a weapon with an actor's ammunition item
+ *
+ * @param {object} actor The actor who holds the weapon
+ * @param {object} weapon The used weapon
+ * @param {object} usedAmmo The Ammo that is used
+ */
+async function reloadAmmo(actor, weapon, usedAmmo) {
+  if (weapon.system.ammo.value === weapon.system.ammo.max) {
+    ChatMessage.create({
+      content: `${game.i18n.format('torgeternity.chatText.ammoFull', { a: weapon.name })}`,
+      speaker: ChatMessage.getSpeaker(),
+    });
+    return;
+  }
+
+  if (!usedAmmo) {
+    // called from the main actor sheet, it's not known what ammo item is used.
+    const ammoArray = [];
+    for (const item of actor.items) {
+      item.type === 'ammunition' && ammoArray.push(item);
+    }
+    if (ammoArray.length === 0) {
+      ChatMessage.create({
+        content: `${game.i18n.localize('torgeternity.chatText.noAmmoPosessing')}`,
+        speaker: ChatMessage.getSpeaker(),
+      });
+      return;
+    }
+
+    if (ammoArray.length === 1) {
+      usedAmmo = ammoArray[0];
+    } else {
+      let dialogContent =
+        '<p>' +
+        game.i18n.localize('torgeternity.dialogWindow.chooseAmmo.maintext') +
+        '</p><form style="margin-bottom: 1rem"><div style="display:flex;flex-direction: column; list-style: none; align-items:center; gap=3px">';
+
+      for (const ammo of ammoArray) {
+        dialogContent += `<span><input id="${ammo.id}" name="chooseAmmoRdb" data-chosen-id="${ammo.id}" type="radio"/>
+      <label for="${ammo.id}">${ammo.name}</label>
+      </span>`;
+      }
+
+      dialogContent += '</form></div>';
+
+      await Dialog.wait({
+        title: game.i18n.localize('torgeternity.dialogWindow.chooseAmmo.windowTitle'),
+        content: dialogContent,
+        buttons: {
+          ok: {
+            label: `${game.i18n.localize('torgeternity.submit.OK')}`,
+            callback: (html) => {
+              const rdbElements = html[0].getElementsByTagName('input');
+              for (const rdb of rdbElements) {
+                if (rdb.checked) {
+                  usedAmmo = actor.items.get(rdb.dataset.chosenId);
+                  return;
+                }
+              }
+            },
+          },
+          abort: {
+            label: `${game.i18n.localize('torgeternity.submit.cancel')}`,
+            callback: () => {
+              return;
+            },
+          },
+        },
+        default: 'ok',
+      });
+    }
+  }
+
+  await weapon.update({ 'system.ammo.value': weapon.system.ammo.max });
+
+  usedAmmo.system.quantity === 1
+    ? await actor.deleteEmbeddedDocuments('Item', [usedAmmo.id])
+    : await usedAmmo.update({ 'system.quantity': usedAmmo.system.quantity - 1 });
+
+  await ChatMessage.create({
+    content: game.i18n.format('torgeternity.chatText.reloaded', { a: weapon.name }),
+    speaker: ChatMessage.getSpeaker(),
+  });
+}
+
+export { reloadAmmo };